requirements.txt
Django==4.1.1
gunicorn
asgiref==3.5.2
django-cors-headers==3.11.0
django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
psycopg2==2.9.1
PyJWT==2.3.0
pytz==2022.1
sqlparse==0.4.2
typing-extensions==4.2.0
django-import-export==2.9.0
gunicorn
django-storages
boto3
Pillow==9.2.0
django-image-optimizer
django-admin-list-filters